Widen query abundance to 64-bit in search/cluster/chimera and the library API
The query abundance was read from fastx_get_abundance() (int64_t) into a
plain int in orient, search, search_exact and sintax, and carried as int
through several internal helpers (search/cluster) and the public library
API (search_session_single, search_batch, chimera_detect_single,
chimera_detect_batch). A query ;size= above INT_MAX therefore wrapped to a
negative value before reaching the size-ratio / abskew filters
(searchcore.cc maxqsize/minsizeratio/maxsizeratio), silently dropping or
misclassifying the query.

Widen the query-abundance reads, the internal helper parameters, and the
library API (query_size scalars to int64_t, query_sizes arrays to
const int64_t *) to match fastx_get_abundance()/db_getabundance() and the
int64_t abundance accumulators. The api_examples are updated to pass
int64_t size arrays.

This changes the public library ABI, so bump the API version 0.1.0 -> 0.2.0.

+++ src/chimera.h
@@ -143,7 +143,7 @@ auto chimera_detect_single(struct chimera_info_s * ci,
                            const char * query_seq,
                            const char * query_head,
                            int query_len,
-                           int query_size,
+                           int64_t query_size,
                            struct chimera_result_s * result) -> int;
 
 /* Convenience: chimera_detect_thread_cleanup(ci) + chimera_session_cleanup().
@@ -165,6 +165,6 @@ auto chimera_detect_cleanup(struct chimera_info_s * ci) -> void;
 auto chimera_detect_batch(const char ** query_seqs,
                           const char ** query_heads,
                           const int * query_lens,
-                          const int * query_sizes,
+                          const int64_t * query_sizes,
                           int query_count,
                           struct chimera_result_s * results) -> void;
